Moved my stuff into a subfolder and set up method to make API call to image API
