+++ AndrewDev/test.py
@@ -1,0 +1,3 @@
+from place import *
+
+print(test)
